Fixes
Simple interface
Removed keys movement
Some code moved to separate files
