+++ package.json
@@ -11,7 +11,7 @@
     "online"
   ],
   "author": "Nick Kosarev <hmbanan666@hotmail.com>",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/hmbanan666/royal-madness-client/issues"
   },
