+++ src/App.css
@@ -83,3 +83,14 @@
         transform: translate3d(-100%, 0%, 0);
     }
 }
+
+.interface {
+    position: absolute;
+    top: 10px;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 2;
+    font-weight: bold;
+    line-height: 1.2;
+}
