+++ src/game/constants.ts
@@ -1,1 +1,2 @@
 export const PIXEL_SIZE = 1;
+export const FPS = 1000 / 60;
